climbing: refactor to static GRADE_TABLE (#391)
